refactor: reorganize map providers
