+++ src/base/static/legacy-libs/binaryajax.js
