+++ src/base/static/legacy-libs/exif.js
