+++ src/base/static/legacy-libs/jquery-1.10.2.js
