+++ src/base/static/libs/maps/mapboxgl-provider.js
@@ -1,6 +1,6 @@
 /* global MAP_PROVIDER_TOKEN */
 
-import AbstractMapFactory from "./abstract-map-factory";
+import AbstractMapFactory from "./abstract-provider";
 import mapboxgl from "mapbox-gl";
 import "mapbox-gl/dist/mapbox-gl.css";
