+++ src/base/templates/base.hbs
@@ -181,9 +181,9 @@
   {{/if}}
   
   <!-- These are obscure or customized, self-hosted. -->
-  <script src="/libs/binaryajax.js"></script>
-  <script src="/libs/exif.js"></script>
-  <script src="/libs/swag.min.js"></script>
+  <script src="/legacy-libs/binaryajax.js"></script>
+  <script src="/legacy-libs/exif.js"></script>
+  <script src="/legacy-libs/swag.min.js"></script>
 
   <!-- Bootstrap site and user information -->
   <script>
